refactor: 동아리원 명단 수정 api

+++ src/main/java/ddingdong/ddingdongBE/domain/club/service/ClubMemberService.java
@@ -3,6 +3,7 @@ package ddingdong.ddingdongBE.domain.club.service;
 import static ddingdong.ddingdongBE.common.exception.ErrorMessage.NO_SUCH_CLUB;
 
 import ddingdong.ddingdongBE.domain.club.controller.dto.request.ClubMemberDto;
+import ddingdong.ddingdongBE.domain.club.controller.dto.request.UpdateClubMemberRequest;
 import ddingdong.ddingdongBE.domain.club.entity.Club;
 import ddingdong.ddingdongBE.domain.club.entity.ClubMember;
 import ddingdong.ddingdongBE.domain.club.repository.ClubMemberRepository;
@@ -10,6 +11,7 @@ import ddingdong.ddingdongBE.domain.club.repository.ClubRepository;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.apache.poi.ss.usermodel.CellType;
@@ -28,20 +30,43 @@ public class ClubMemberService {
     private final ClubRepository clubRepository;
     private final ClubMemberRepository clubMemberRepository;
 
-    public void updateClubMembers(Long userId, MultipartFile clubMemberListFile) {
-        isExcelFile(clubMemberListFile);
+    public void updateClubMembers(Long userId, UpdateClubMemberRequest request,
+                                  Optional<MultipartFile> clubMemberListFile) {
+        if (clubMemberListFile.isPresent()) {
+            MultipartFile file = clubMemberListFile.get();
+            isExcelFile(file);
+
+            Club club = clubRepository.findByUserId(userId)
+                    .orElseThrow(() -> new IllegalArgumentException(NO_SUCH_CLUB.getText()));
+            List<ClubMemberDto> requestedClubMemberDtos = parsingClubMemberListFile(file);
+            List<ClubMember> requestedClubMembers = requestedClubMemberDtos.stream()
+                    .map(clubMemberDto -> clubMemberDto.toEntity(club))
+                    .toList();
+
+            clubMemberRepository.deleteAllByClub(club);
+            clubMemberRepository.saveAll(requestedClubMembers);
+            return;
+        }
 
         Club club = clubRepository.findByUserId(userId)
                 .orElseThrow(() -> new IllegalArgumentException(NO_SUCH_CLUB.getText()));
-        List<ClubMemberDto> requestedClubMemberDtos = parsingClubMemberListFile(
-                clubMemberListFile);
-        List<ClubMember> requestedClubMembers = requestedClubMemberDtos.stream()
+
+        List<Long> memberIds = clubMemberRepository.findClubMembersByClubId(club.getId())
+                .stream()
+                .map(ClubMember::getId)
+                .toList();
+
+        List<ClubMember> requestedClubMembers = request.getClubMemberList().stream()
                 .map(clubMemberDto -> clubMemberDto.toEntity(club))
                 .toList();
 
-        clubMemberRepository.deleteAllByClub(club);
+        if (!memberIds.isEmpty()) {
+            clubMemberRepository.deleteAllById(memberIds);
+        }
+
         clubMemberRepository.saveAll(requestedClubMembers);
     }
+
 
     private void isExcelFile(MultipartFile file) {
         String fileName = file.getOriginalFilename();
@@ -61,7 +86,7 @@ public class ClubMemberService {
                 }
             }
         } catch (IOException e) {
-            throw new IllegalArgumentException(e);
+            throw new IllegalArgumentException("올바른 엑셀 양식을 사용해주세요.");
         }
         return requestedClubMemberDtos;
     }
